Former-commit-id: 8d8541da25bb5ddcb27b23744132ae64fd4a9b22 [formerly 046e5256840c7786bada079e96a0a081730a4b9d]
Former-commit-id: 5a99d2af25faa687ea8554e0c6f6588155f4befc

+++ openquake/calculators/tests/__init__.py
@@ -165,7 +165,11 @@ class CalculatorTestCase(unittest.TestCase):
 
     def run(self, result=None):
         res = super(CalculatorTestCase, self).run(result)
+        if res is not None:  # for Python 3
+            issues = len(res.errors) + len(res.failures)
+        else:
+            issues = 0  # this is bad, but Python 2 will die soon or later
         # remove temporary dir only for success
-        if self.edir:
+        if self.edir and not issues:
             shutil.rmtree(self.edir)
         return res
